fix message in like comment & add optionalAuthMiddleware

+++ src/services/comment-like.service.js
@@ -1,16 +1,20 @@
 import CommentLike from "../models/comment-like.model.js"
+import Comment from "../models/comment.model.js"
+import { NotFoundError } from "../handler/error-response.js"
 
 class CommentLikeService {
-  toggleLike = async (userId, commentId) => {
-    const existingLike = await CommentLike.findOne({ user: userId, comment: commentId })
+  toggleLike = async (commentId, userId) => {
+    const comment = await Comment.findById(commentId)
+    if (!comment) throw new NotFoundError('Comment not found')
 
+    const existingLike = await CommentLike.findOne({ comment: commentId, user: userId })
     if (existingLike) {
       await existingLike.deleteOne()
-      return { message: 'Like removed' }
+      return { liked: false }
     } else {
       const newLike = new CommentLike({ comment: commentId, user: userId })
       await newLike.save()
-      return { message: 'Like added' }
+      return { liked: true }
     }
   }
 
@@ -18,7 +22,7 @@ class CommentLikeService {
     return await CommentLike.countDocuments({ comment: commentId })
   }
 
-  isLikedByCurrentUser = async (commentId, userId) => {
+  comment_isLikedByCurrentUser = async (commentId, userId) => {
     if (!userId) return false
     const like = await CommentLike.findOne({ comment: commentId, user: userId })
     return !!like
